Fix warnings, organize imports

+++ src/main/scala/lambdas/challenge/CharacterReaderInputStream.scala
@@ -1,14 +1,14 @@
 package lambdas.challenge
 
-import java.io.InputStream
 import java.io.EOFException
+import java.io.InputStream
 
 /** An adapter to turn a [[lambdas.challenge.CharacterReader]] into an [[java.io.InputStream]]. */
 class CharacterReaderInputStream(reader: CharacterReader) extends InputStream {
 
   override def read(): Int = {
     try {
-      reader.nextCharacter()
+      reader.nextCharacter().toInt
     } catch {
       case _: EOFException => -1
     }
